Split test_zinc_compile_integration into two smaller tests
